+++ lib/src/modules/elss/common_providers/user_session_provider.dart
@@ -18,11 +18,11 @@ class UserSessionProvider extends CshChangeNotifier {
     switch (loginTypeEnum) {
       case LoginTypes.trcLogin:
         return ElssService.trcLogout();
-      case LoginTypes.qcLogin:
-        return ElssService.qcLogout();
       case LoginTypes.rmsLogin:
       case LoginTypes.shipexLogin:
         return ElssService.consoleLogout();
+      case LoginTypes.qcLogin:
+        return Stream.value(LogoutResponse(1, "", ""));
     }
   }